# Optimize N-Quads escape/unescape.