--- a/lib/NQuads.js
+++ b/lib/NQuads.js
@@ -363,6 +363,9 @@ const _escapeRegex = /["\\\n\r]/g;
  * Escape string to N-Quads literal
  */
 function _escape(s) {
+  if(!_escapeRegex.test(s)) {
+    return s;
+  }
   return s.replace(_escapeRegex, function(match) {
     switch(match) {
       case '"': return '\\"';
@@ -379,6 +382,9 @@ const _unescapeRegex =
  * Unescape N-Quads literal to string
  */
 function _unescape(s) {
+  if(!_unescapeRegex.test(s)) {
+    return s;
+  }
   return s.replace(_unescapeRegex, function(match, code, u, U) {
     if(code) {
       switch(code) {