- Run regex test before doing a replace call.
- Performance difference depends on data and how often escape/unescape would need to be called. A benchmark test data showed ~3-5% overall improvement.

The benchmark here shows improvement, but in the case where data is always being escaped, this patch would be slower.  We could add flags to control this, but for now making an assumption overall this will be an improvement.

# test-then-replace

## Comparison
| Test                 |   'base' | 'new' | 'new' |
| :------------------- | -------: | ----: | ----: |
| #manifest#block-1    | 11111.08 | 5.62% | 5.55% |
| #manifest#block-2    |  2719.06 | 3.64% | 5.17% |
| #manifest#block-10   |   604.62 | 4.28% | 4.07% |
| #manifest#block-100  |    59.56 | 0.96% | 3.15% |
| #manifest#block-1000 |     5.50 | 3.04% | 4.00% |

> base ops/s and relative difference (higher is better)

## Environment
| Key             | Values                                   |
| --------------- | ---------------------------------------- |
| Label           | 'base', 'new'                            |
| Arch            | x64                                      |
| CPU             | Intel(R) Core(TM) i7-4790K CPU @ 4.00GHz |
| CPUs            | 8                                        |
| Platform        | linux                                    |
| Runtime         | Node.js                                  |
| Runtime Version | v16.19.1                                 |
| Comment         | 'test-then-replace'                      |